terraform/openstack: Enable predefined floating IP.
The implementation used to define k8s masters and nodes in detail
doesn't currently have a way to assign existing floating IPs.

This commit allows this assignment for masters through another option in
the `k8s_masters` variable called `floating_ip_address`, the existence
of which will cause the module to assign the floating IP to the master.