[openstack/terraform] Config changes to allow GlusterFS deployment.
The resources in the terraform script that create GlusterFS VMs accept
image UUIDs in one place and names in another. This causes either a
`name wrong` or a `uuid wrong` error. I made those identical, so the UUID
would be used properly.

Also, JetStream2 does not seem to accept the volumev2 API even though it
says so on their `API Access` page on the console. So I changed it to
V3.